Gamepad support and Pet Commands fix
ADDED GAMEPAD SUPPORT: To use please unmap your confirm key in FFXIPadConfig and run the included .ahk file before starting the game. This .ahk was made with DS4 Tool and a PS4 controller in mind and may need modification for your setup. Also be sure to change all my character names to your characters names.

Fixed Pet Commands not showing up as abilities and added item as a possible type of command.